+++ R/tf.R
@@ -42,6 +42,7 @@
 #' @return An updated version of `recipe` with the new step added
 #'  to the sequence of existing steps (if any).
 #' @examples
+#' if (requireNamespace("text2vec", quietly = TRUE)) {
 #' \donttest{
 #' library(recipes)
 #' 
@@ -58,6 +59,7 @@
 #' 
 #' tidy(okc_rec, number = 2)
 #' tidy(okc_obj, number = 2)
+#' }
 #' }
 #' @export
 #' @details