+++ R/word2vec.R
@@ -37,6 +37,7 @@
 #'  to the sequence of existing steps (if any).
 #' @source \url{https://arxiv.org/abs/1301.3781}
 #' @examples
+#' if (requireNamespace("text2vec", quietly = TRUE)) {
 #' library(recipes)
 #'
 #' data(okc_text)
@@ -73,7 +74,7 @@
 #'   prep() %>%
 #'   juice() %>%
 #'   slice(1:2)
-#'
+#' }
 #' @export
 #'
 #' @importFrom recipes add_step step terms_select sel2char ellipse_check