+++ tests/testthat/test-word2vec.R
@@ -8,6 +8,7 @@ n_rows <- 100
 rec <- recipe(~ essay0 + essay1, data = okc_text[seq_len(n_rows), ])
 
 test_that("step_word2vec works as intended", {
+  skip_if_not_installed("text2vec")
   n_top <- 10
   rec1 <- rec %>%
     step_word2vec(essay0, num_topics = n_top)
@@ -22,6 +23,7 @@ test_that("step_word2vec works as intended", {
 })
 
 test_that("step_word2vec works with num_topics argument", {
+  skip_if_not_installed("text2vec")
   n_top <- 100
   rec1 <- rec %>%
     step_word2vec(essay0, num_topics = n_top)
@@ -33,6 +35,7 @@ test_that("step_word2vec works with num_topics argument", {
 })
 
 test_that("printing", {
+  skip_if_not_installed("text2vec")
   rec <- rec %>%
     step_word2vec(essay0)