+++ NEWS.md
@@ -1,5 +1,7 @@
 # textrecipes (development version)
 
+* `step_word_embeddings()` has been added. Allows for use of pre-trained word embeddings to convert token columns to vectors in a high-dimensional "meaning" space. (@jonthegeek, #20)
+
 # textrecipes 0.0.2
 
 * Custom stemming function can now be used in step_stem using the custom_stemmer argument.