+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## Changed behaviour
 
+* text2vec have been changed from Imports to Suggests.
 * step_tfidf() calculations are slightly changed due to flaw in original implementation https://github.com/dselivanov/text2vec/issues/280.
 
 # textrecipes 0.0.2