+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Changed behaviour
 
 * text2vec have been changed from Imports to Suggests.
+* textfeatures have been changed from Imports to Suggests.
 * step_tfidf() calculations are slightly changed due to flaw in original implementation https://github.com/dselivanov/text2vec/issues/280.
 
 # textrecipes 0.0.2