+++ NEWS.md
@@ -4,7 +4,7 @@
 * `step_word_embeddings()` has been added. Allows for use of pre-trained word embeddings to convert token columns to vectors in a high-dimensional "meaning" space. (@jonthegeek, #20)
 * text2vec have been changed from Imports to Suggests.
 * textfeatures have been changed from Imports to Suggests.
-* step_tfidf() calculations are slightly changed due to flaw in original implementation https://github.com/dselivanov/text2vec/issues/280.
+* `step_tfidf()` calculations are slightly changed due to flaw in original implementation https://github.com/dselivanov/text2vec/issues/280.
 
 # textrecipes 0.0.2